lms_backend/requirements.txt
bcrypt==4.1.2
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==3.1.5
